+++ config.go
@@ -2,7 +2,7 @@ package selfupdate
 
 // Config represents the configuration of self-update.
 type Config struct {
-	// Source where to load the releases from (example: GitHubSource)
+	// Source where to load the releases from (example: GitHubSource).
 	Source Source
 	// Validator represents types which enable additional validation of downloaded release.
 	Validator Validator
@@ -10,14 +10,16 @@ type Config struct {
 	// An asset is selected if it matches any of those, in addition to the regular tag, os, arch, extensions.
 	// Please make sure that your filter(s) uniquely match an asset.
 	Filters []string
-	// OS is set to the value of runtime.GOOS by default, but you can force another value here
+	// OS is set to the value of runtime.GOOS by default, but you can force another value here.
 	OS string
-	// Arch is set to the value of runtime.GOARCH by default, but you can force another value here
+	// Arch is set to the value of runtime.GOARCH by default, but you can force another value here.
 	Arch string
-	// Arm 32bits version. Valid values are 0 (unknown), 5, 6 or 7. Default is detected value (if any)
+	// Arm 32bits version. Valid values are 0 (unknown), 5, 6 or 7. Default is detected value (if available).
 	Arm uint8
-	// Draft permits an upgrade to a "draft" version (default to false)
+	// Arch name to use when using a universal binary (macOS only). Default to none.
+	UniversalArch string
+	// Draft permits an upgrade to a "draft" version (default to false).
 	Draft bool
-	// Prerelease permits an upgrade to a "pre-release" version (default to false)
+	// Prerelease permits an upgrade to a "pre-release" version (default to false).
 	Prerelease bool
 }
